requirements.txt
apscheduler==3.10.4
croniter==2.0.2
pyyaml==6.0.1
requests==2.31.0
